+++ CHANGES.md
@@ -6,6 +6,10 @@ Version 0.7.0
 
 To be released.
 
+ -  Added `Bencodex.Types.Binary.FromHex()` static method.  [[#69]]
+
+[#69]: https://github.com/planetarium/bencodex.net/pull/69
+
 
 Version 0.6.0
 -------------